#1 PR changes (namings, exceptions)
