+++ WhoIsThat/WhoIsThat/Views/HomePage.xaml
@@ -5,11 +5,11 @@
              Padding="60, 5, 60, 30">
     <ContentPage.Content>
         <StackLayout>
-            <Button Text="Take a photo!"
+            <Button Text="Take Photo!"
                     TextColor="White"
                     BackgroundColor="Black"
                     FontSize="30"
-                    Clicked="takePicture"></Button>
+                    Clicked="TakePicture"></Button>
 
             <Image x:Name="takenPicture"></Image>
         </StackLayout>
