+++ WhoIsThat/WhoIsThat/Views/MainPage.xaml.cs
