+++ WhoIsThat/WhoIsThat/Views/RegistrationPage.xaml
