+++ WhoIsThat/WhoIsThat/Views/RegistrationPage.xaml.cs
